Switch to JFlex 1.8.2 for real and fix compat bugs
My local Netbeans installation still had a copy of JFlex 1.6 that was overriding the version in /lib. Removing that and letting JFlex 1.8.2 run on the .flex files caused some minor issues with long-to-int conversion, which this commit fixes.

+++ EugSyntax/src/eug/syntax/EUGDocument.java
@@ -112,11 +112,11 @@ public class EUGDocument extends PlainDocument {
             p0 = start;
         }
         
-        public int getStartOffset() {
+        public long getStartOffset() {
             return p0 + scanner.getTokenStart();
         }
         
-        public int getEndOffset() {
+        public long getEndOffset() {
             return p0 + scanner.getTokenEnd();
         }
